[#2245] Rework Locale selector UI
