+++ client/src/components/workspace-nav/user-menu/LocaleSelector.scss
@@ -1,0 +1,12 @@
+a.LocaleSelectorItem {
+  margin-right: 1em;
+  padding: 0.4em;
+  border-radius: 3px;
+}
+a.LocaleSelectorItem:hover {
+  background-color: #f1f2f2;
+}
+
+a.LocaleSelectorItemSelected {
+  background-color: #f1f2f2;
+}
